loan_underwriting/requirements.txt
gradio>=4.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11 